pyproject.toml
[project]
name = "A-multi-agent-research-assistant-using-autogen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "pandas>=2.3.3",
    "pyautogen>=0.10.0",
    "python-dotenv>=1.2.1",
]
